chore: remove simulate.ps1 and CLAUDE.md reference
Script is no longer useful; e2e tests and the proxy itself cover
traffic simulation needs.

+++ sevorix-core/src/pty_multiplexer.rs
@@ -315,7 +315,8 @@ impl BashPty {
                 let _ = setsid();
 
                 // Set controlling terminal
-                let _ = unsafe { libc::ioctl(pty_pair.slave_fd, libc::TIOCSCTTY, 0) };
+                let _ =
+                    unsafe { libc::ioctl(pty_pair.slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0) };
 
                 // Dup slave to stdin/stdout/stderr
                 let _ = dup2(pty_pair.slave_fd, 0);
